Assume our UNPAIR expansion to be valid
