+++ packages/taquito-michel-codec/src/taquito-michel-codec.ts
@@ -5,5 +5,6 @@ export * from "./michelson-validator";
 export * from "./michelson-types";
 export * from "./michelson-typecheck";
 export * from "./michelson-contract";
-export { MichelsonError, isMichelsonError, formatStack, formatError, traceDumpFunc } from "./utils";
+export * from "./formatters";
+export { MichelsonError, isMichelsonError } from "./utils";
 export { MacroError } from "./macros";
